Mutliple changes
A few bits and pieces, some inportant, others less so. Change list in no
particular order.

• Added License.txt
• Updated nuspec files
• Renamed test namespace from JsonFileDefinitionParser (former name) to
JsonFileCommandReader (better name).
• Added parameterless Execute methods for sync/async (and minimal test
coverage for both)
• Moved "NamespaceB" in Drapper.Test.Settings.json from its misleading
area of the test config into a PathSetting of the root namespace (as it
should have been from the start).
• Updated a few of the helper classes to use single instance of their
respective _commander variables.
• Set build to release with /optimize
• Added MethodImpl attribues on the DbCommander in attempt to fix Issue
#1
• Temporarily suspended support for ExecuteAsync without Type parameter.
The Type parameter is no mandatory of ExecuteAsync calls. Trying this
out at the moment. Needs to applied consistently if I choose to go down
this route.
• Updated file headers all over the show.